Fix errors
This took me so long to do. I am drained. Please help me. Im
making so many retarded mistakes and getting so friggin
confused.
Basically, I fixed most of the issues as far as I know. The
data is being stored in the localStorage and is being fetched
back at the start of the program successfully. The set functions
for the Tasks are also being saved to the actual database.