Update database structure
It was an array of all the tasks objects prior to this commit.
Now the database is basically an object with key value pairs of
'project name' : [array of the project's tasks]
A sorter is also implemented to ensure newly created tasks are
sorted into their respective projects at creation.
The default task array has the key (from key:value) of 'none' and
the unsorted tasks are temporarily stored in an array with the
key of 'unsorted'.